Split JDomUtils into smaller file chunks
* JDomCleanupHelper: Methods to cleanup elements before writing the pom
* JDomContentHelper: Utility methods to work with content
* JDomUtils        : Manipulation the pom (add, insert, rewrite, and remove)
